Stop packaging Libtool `.la` files on Windows.
Pursuant to https://github.com/conda-forge/staged-recipes/issues/673 and
https://github.com/conda-forge/libxcb-feedstock/pull/9, we're going to remove
these files from our packages. Eventually conda-build will do this
automatically for us, but for now we take care of it manually. We need to
delete the files both before the build (because inconsistent complements of
`.la` files can break the build) and after (because we may have installed new
ones).

We also switch to using `--disable-static` to avoid packaging static
libraries. I had avoided this because on Windows you need to distribute
`.dll.a` files for libraries to link to one another, but it turns out that The
Right Thing happens anyway. (This isn't so surprising, but I just hadn't done
the work to actually check this before.)